♻️ move graphql definitions in dedicated module
