+++ marketplace-core/src/graphql/query.rs
@@ -1,0 +1,14 @@
+use juniper::graphql_object;
+
+pub struct Query;
+
+#[graphql_object]
+impl Query {
+	pub fn new() -> Self {
+		Self {}
+	}
+
+	pub fn hello(&self) -> &str {
+		"Couscous!"
+	}
+}
